Full OIDC providers implementation
You must go to it in the browser with this url since I didnt setup
proper redirects.

https://habits.example.com/auth/<ID>/login?return=/auth/get_api_token

1. Go to url above (replace your domain and ID) and login
2. take return string and save to JWT_TOKEN env var
3. curl https://habits.example.com/habits/ -H "Authorization: Bearer ${JWT_TOKEN}"

```bash
$ curl https://habits.yapplesauce.com/api/habits -H "Authorization: Bearer ${JWT_TOKEN}"
{"habits":[]}
```

Properly isolate issuers in routes

Now the URLs for the auth are:
  - /auth/<ID>/login
  - /auth/<ID>/callback
  - /auth/<ID>/logout

What we need is a /auth/login page that will let you select your
provider. And then that page redirects them to the unique route with the
stable ULID id.

I have also solved some of the cookie situatation. Though we might still
be able to switch to pure session ID by moving the remaining cookie to
the main server