Add slog for log levels
In my research, `slog` is the choice to use. It isn't an external dep,
and thats all I need to really hear.

It can also do console writes simultaneously with writing to a log as a
configuration step. Not sure I expect a log file to really be useful to
this project, but it is easy to add and configure if you want.

Adds a new `log_level` config option that defaults to `INFO`